+++ data.php
@@ -1,0 +1,42 @@
+<?php 
+
+$project_cats=["Все","Входящие","Учеба","Работа","Домашние дела","Авто"];
+
+$array_tasks = [
+0=> [
+'task_name' => 'Собеседование в IT компании',
+'date_of_deadline' => '01.06.2018',
+'task_category'=>'Работа',
+'task_done'=>false
+],
+1=> [
+'task_name' => 'Выполнить тестовое задание',
+'date_of_deadline' => '20.05.2018',
+'task_category'=>'Работа',
+'task_done'=>false
+],
+2=> [
+'task_name' => 'Сделать задание первого раздела',
+'date_of_deadline' => '24.01.2018',
+'task_category'=>'Учеба',
+'task_done'=>true
+],
+3=> [
+'task_name' => 'Встреча с другом',
+'date_of_deadline' => '22.04.2018',
+'task_category'=>'Входящие',
+'task_done'=>false
+],
+4=> [
+'task_name' => 'Купить корм для кота',
+'date_of_deadline' => null,
+'task_category'=>'Домашние дела',
+'task_done'=>false
+],
+5=> [
+'task_name' => 'Заказать пиццу',
+'date_of_deadline' => null,
+'task_category'=>'Домашние дела',
+'task_done'=>false
+]
+];  ?>
